requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
easyocr==1.7.1
opencv-python==4.9.0.80
numpy==1.24.3
Pillow==10.2.0
Werkzeug==3.0.1